Fix compilation warning on macOS

+++ src/sycomore/Quantity.h
@@ -183,7 +183,7 @@ SYCOMORE_API Quantity operator/(Quantity l, Quantity const & r);
 
 /// @brief Division
 template<typename T, typename std::enable_if<std::is_arithmetic<T>::value, int>::type=0>
-Quantity operator/(Quantity q, T s) { q /= s; return q; }
+Quantity operator/(Quantity q, T s) { q /= static_cast<double>(s); return q; }
 
 /// @brief Division
 template<typename T, typename std::enable_if<std::is_arithmetic<T>::value, int>::type=0>
